Port test refactoring for enumerating all services
Upstream botocore PR: https://github.com/boto/botocore/pull/2986
