+++ tests/__init__.py
@@ -56,9 +56,10 @@ from prompt_toolkit.input.ansi_escape_sequences import REVERSE_ANSI_SEQUENCES
 # Botocore testing utilities that we want to preserve import statements for
 # in botocore specific tests.
 from tests.utils.botocore import (
-    assert_url_equal, create_session, random_chars, temporary_file, BaseEnvVar,
-    BaseSessionTest, BaseClientDriverTest, StubbedSession, ClientHTTPStubber,
-    SessionHTTPStubber, IntegerRefresher, FreezeTime,
+    assert_url_equal, create_session, random_chars, temporary_file,
+    patch_load_service_model, ALL_SERVICES, BaseEnvVar, BaseSessionTest,
+    BaseClientDriverTest, StubbedSession, ClientHTTPStubber, SessionHTTPStubber,
+    IntegerRefresher, FreezeTime,
 )
 # S3transfer testing utilities that we want to preserve import statements for
 # in s3transfer specific tests.
